requirements.txt
Flask
gunicorn
mysql-connector-python
pandas
numpy
matplotlib
seaborn
scikit-learn
Werkzeug
lxml
python-dotenv